+++ benchmark/merge-path/merge_path_spmv.h
@@ -6,7 +6,7 @@
 #include "merge_path_config.h"
 
 template <int REDUCTION_ALGORITHM = Binary, int UPDATE_ALGORITHM = SingleBlock>
-void merge_path_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+void merge_path_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                      const csr_desc<int, double> d_csr_desc, const double *x, double *y, BenchmarkTime *bmt);
 
 #endif // SPMV_ACC_BENCHMARK_MERGE_PATH_SPMV_H
