+++ src/acc/hip-light/light_spmv.cpp
@@ -13,7 +13,7 @@
 #define LIGHT_KERNEL_CALLER(N)                                                                                         \
   ((spmv_light_kernel<N, __WF_SIZE__, double>) <<<256,256>>> (m, alpha, beta, hip_row_counter, rowptr, colindex, value, x, y))
 
-void light_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> d_csr_desc,
+void light_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> d_csr_desc,
                        const double *x, double *y) {
   VAR_FROM_CSR_DESC(d_csr_desc);
