+++ src/acc/hip-line-enhance/line_enhance_spmv.h
@@ -7,13 +7,13 @@
 
 #include "api/types.h"
 
-void line_enhance_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> d_csr_desc,
+void line_enhance_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> d_csr_desc,
                               const double *x, double *y);
 
 /**
  * set kernel template parameters adaptively by matrix characteristic.
  */
-void adaptive_enhance_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> d_csr_desc,
+void adaptive_enhance_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> d_csr_desc,
                                   const double *x, double *y);
 
 #endif // SPMV_ACC_LINE_ENHANCE_SPMV_H
